feat: README per plugin
