+++ package.json
@@ -40,7 +40,6 @@
     "eslint-plugin-n": "17.24.0",
     "knip": "6.11.0",
     "prettier": "3.8.3",
-    "tsdown": "catalog:dev",
     "typescript": "6.0.3",
     "typescript-eslint": "catalog:dev"
   }
