+++ packages/utils/src/index.ts
@@ -1,17 +1,11 @@
-export {
-  buildZodChainRemoveMethodFix,
-} from './build-zod-chain-remove-method-fix.js';
-export {
-  buildZodChainReplacementFix,
-} from './build-zod-chain-replacement-fix.js';
+export { buildZodChainRemoveMethodFix } from './build-zod-chain-remove-method-fix.js';
+export { buildZodChainReplacementFix } from './build-zod-chain-replacement-fix.js';
 export {
   detectZodSchemaRootNode,
   isZodNumberSchemaCallExpression,
 } from './detect-zod-schema-root-node.js';
 export type { DetectResult } from './detect-zod-schema-root-node.js';
-export {
-  findParentSchemaMatchingCondition,
-} from './find-parent-schema-matching-condition.js';
+export { findParentSchemaMatchingCondition } from './find-parent-schema-matching-condition.js';
 export {
   isZodImportSource,
   ZOD_IMPORT_SOURCES,
