+++ plugins/eslint-plugin-zod-mini/docs/rules/consistent-import-source.md
@@ -1,0 +1,17 @@
+# zod-mini/consistent-import-source
+
+📝 Enforce consistent source from Zod Mini imports.
+
+💡 This rule is manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name      | Description                             | Type     |
+| :-------- | :-------------------------------------- | :------- |
+| `sources` | An array of allowed Zod import sources. | String[] |
+
+<!-- end auto-generated rule options list -->
