+++ plugins/eslint-plugin-zod-mini/docs/rules/consistent-import.md
@@ -1,0 +1,19 @@
+# zod-mini/consistent-import
+
+📝 Enforce a consistent import style for Zod Mini.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name     | Description                                      | Type   | Choices              |
+| :------- | :----------------------------------------------- | :----- | :------------------- |
+| `syntax` | Specifies the import syntax to use for Zod Mini. | String | `namespace`, `named` |
+
+<!-- end auto-generated rule options list -->
