+++ plugins/eslint-plugin-zod-mini/docs/rules/consistent-object-schema-type.md
@@ -1,0 +1,17 @@
+# zod-mini/consistent-object-schema-type
+
+📝 Enforce consistent usage of Zod Mini schema methods.
+
+💡 This rule is manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name    | Description                              | Type     |
+| :------ | :--------------------------------------- | :------- |
+| `allow` | Decides which object methods are allowed | String[] |
+
+<!-- end auto-generated rule options list -->
