+++ plugins/eslint-plugin-zod-mini/docs/rules/consistent-schema-output-type-style.md
@@ -1,0 +1,17 @@
+# zod-mini/consistent-schema-output-type-style
+
+📝 Enforce consistent use of z.infer or z.output for schema type inference.
+
+🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name    | Description                                          | Type   | Choices           |
+| :------ | :--------------------------------------------------- | :----- | :---------------- |
+| `style` | Decides which style to use for schema type inference | String | `infer`, `output` |
+
+<!-- end auto-generated rule options list -->
