+++ plugins/eslint-plugin-zod-mini/docs/rules/consistent-schema-var-name.md
@@ -1,0 +1,18 @@
+# zod-mini/consistent-schema-var-name
+
+📝 Enforce a consistent naming convention for Zod Mini schema variables.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name     | Description                                       | Type   |
+| :------- | :------------------------------------------------ | :----- |
+| `after`  | The required suffix for Zod Mini schema variables | String |
+| `before` | The required prefix for Zod Mini schema variables | String |
+
+<!-- end auto-generated rule options list -->
