+++ plugins/eslint-plugin-zod-mini/docs/rules/no-any-schema.md
@@ -1,0 +1,9 @@
+# zod-mini/no-any-schema
+
+📝 Disallow usage of `z.any()` in Zod Mini schemas.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+💡 This rule is manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
+
+<!-- end auto-generated rule header -->
