+++ plugins/eslint-plugin-zod-mini/docs/rules/no-empty-custom-schema.md
@@ -1,0 +1,7 @@
+# zod-mini/no-empty-custom-schema
+
+📝 Disallow usage of `z.custom()` without arguments.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+<!-- end auto-generated rule header -->
