+++ plugins/eslint-plugin-zod-mini/docs/rules/no-unknown-schema.md
@@ -1,0 +1,5 @@
+# zod-mini/no-unknown-schema
+
+📝 Disallow usage of `z.unknown()` in Zod Mini schemas.
+
+<!-- end auto-generated rule header -->
