+++ plugins/eslint-plugin-zod-mini/docs/rules/prefer-meta.md
@@ -1,0 +1,9 @@
+# zod-mini/prefer-meta
+
+📝 Enforce usage of `z.meta()` over `z.describe()`.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
+
+<!-- end auto-generated rule header -->
