+++ plugins/eslint-plugin-zod-mini/docs/rules/prefer-namespace-import.md
@@ -1,0 +1,9 @@
+# zod-mini/prefer-namespace-import
+
+📝 Enforce importing zod/mini as a namespace import (`import * as z from 'zod/mini'`).
+
+❌ This rule is deprecated. Use `zod-mini/consistent-import` with `{ syntax: 'namespace' }`.
+
+🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
+
+<!-- end auto-generated rule header -->
