+++ plugins/eslint-plugin-zod-mini/docs/rules/require-brand-type-parameter.md
@@ -1,0 +1,9 @@
+# zod-mini/require-brand-type-parameter
+
+📝 Require type parameter on `.brand()` functions.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+💡 This rule is manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
+
+<!-- end auto-generated rule header -->
