+++ plugins/eslint-plugin-zod-mini/docs/rules/require-error-message.md
@@ -1,0 +1,9 @@
+# zod-mini/require-error-message
+
+📝 Enforce that custom refinements include an error message.
+
+💼 This rule is enabled in the ✅ `recommended` config.
+
+🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
+
+<!-- end auto-generated rule header -->
