+++ plugins/eslint-plugin-zod-mini/docs/rules/require-schema-suffix.md
@@ -1,0 +1,17 @@
+# zod-mini/require-schema-suffix
+
+📝 Require schema suffix when declaring a Zod Mini schema.
+
+❌ This rule is deprecated. Use `zod-mini/consistent-schema-var-name`.
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name     | Description                                       | Type   |
+| :------- | :------------------------------------------------ | :----- |
+| `suffix` | The required suffix for Zod Mini schema variables | String |
+
+<!-- end auto-generated rule options list -->
