+++ plugins/eslint-plugin-zod-mini/docs/rules/schema-error-property-style.md
@@ -1,0 +1,16 @@
+# zod-mini/schema-error-property-style
+
+📝 Enforce consistent style for error messages in Zod Mini schema validation (using ESQuery patterns).
+
+<!-- end auto-generated rule header -->
+
+## Options
+
+<!-- begin auto-generated rule options list -->
+
+| Name       | Description                                                   | Type   |
+| :--------- | :------------------------------------------------------------ | :----- |
+| `example`  | Example code to help the user understand the required pattern | String |
+| `selector` | An ESQuery string to match the required pattern               | String |
+
+<!-- end auto-generated rule options list -->
