+++ plugins/eslint-plugin-zod-mini/src/meta.ts
@@ -4,11 +4,7 @@ const packageJSON = JSON.parse(
   readFileSync(new URL('../package.json', import.meta.url), 'utf8'),
 ) as { name: string; version: string; homepage: string };
 
-const {
-  name: PLUGIN_NAME,
-  version: PLUGIN_VERSION,
-  homepage,
-} = packageJSON;
+const { name: PLUGIN_NAME, version: PLUGIN_VERSION, homepage } = packageJSON;
 
 export { PLUGIN_NAME, PLUGIN_VERSION };
