+++ plugins/eslint-plugin-zod-mini/src/rules/prefer-meta.ts
@@ -26,8 +26,11 @@ export const preferMeta = createZodMiniPluginRule({
   },
   defaultOptions: [],
   create(context) {
-    const { importDeclarationListener, isZodNamespace, getNamedImportOriginal } =
-      trackZodSchemaImports();
+    const {
+      importDeclarationListener,
+      isZodNamespace,
+      getNamedImportOriginal,
+    } = trackZodSchemaImports();
 
     return {
       ImportDeclaration: importDeclarationListener,
