+++ plugins/eslint-plugin-zod/docs/rules/no-any-schema.md
@@ -2,7 +2,7 @@
 
 📝 Disallow usage of `z.any()` in Zod schemas.
 
-💼 This rule is enabled in the following configs: ✅ `recommended`, ✔️ `recommendedMini`.
+💼 This rule is enabled in the ✅ `recommended` config.
 
 💡 This rule is manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
