+++ plugins/eslint-plugin-zod/docs/rules/require-brand-type-parameter.md
@@ -2,7 +2,7 @@
 
 📝 Require type parameter on `.brand()` functions.
 
-💼 This rule is enabled in the following configs: ✅ `recommended`, ✔️ `recommendedMini`.
+💼 This rule is enabled in the ✅ `recommended` config.
 
 💡 This rule is manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
