+++ plugins/eslint-plugin-zod/docs/rules/require-error-message.md
@@ -2,7 +2,7 @@
 
 📝 Enforce that custom refinements include an error message.
 
-💼 This rule is enabled in the following configs: ✅ `recommended`, ✔️ `recommendedMini`.
+💼 This rule is enabled in the ✅ `recommended` config.
 
 🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
