+++ plugins/eslint-plugin-zod/package.json
@@ -14,7 +14,6 @@
   "module": "./dist/index.mjs",
   "exports": {
     ".": {
-      "@eslint-zod/source": "./src/index.ts",
       "import": "./dist/index.mjs",
       "require": "./dist/index.cjs"
     },
@@ -69,10 +68,12 @@
     }
   },
   "devDependencies": {
+    "@marcalexiei/prettier-config": "catalog:dev",
     "@types/esquery": "1.5.4",
     "@typescript-eslint/rule-tester": "catalog:dev",
     "dedent": "catalog:dev",
     "eslint-doc-generator": "catalog:dev",
+    "prettier": "catalog:dev",
     "tsdown": "catalog:dev",
     "vitest": "catalog:dev"
   }
