+++ plugins/eslint-plugin-zod/src/index.ts
@@ -50,7 +50,6 @@ const eslintPluginZod = {
     version: PLUGIN_VERSION,
   },
   rules: {
-     
     'array-style': arrayStyle,
     'consistent-import-source': consistentImportSource,
     'consistent-import': consistentImport,
@@ -81,7 +80,6 @@ const eslintPluginZod = {
     'require-error-message': requireErrorMessage,
     'require-schema-suffix': requireSchemaSuffix,
     'schema-error-property-style': schemaErrorPropertyStyle,
-     
   } as unknown as Record<string, Rule.RuleModule>,
 } satisfies ESLint.Plugin as CompatiblePlugin;
 
@@ -96,7 +94,6 @@ const baseConfig = {
 const recommendedConfig = {
   ...baseConfig,
   rules: {
-     
     'zod/array-style': 'error',
     'zod/consistent-import': 'error',
     'zod/consistent-schema-var-name': 'error',
@@ -118,20 +115,6 @@ const recommendedConfig = {
     'zod/prefer-trim-before-string-length-checks': 'error',
     'zod/require-brand-type-parameter': 'error',
     'zod/require-error-message': 'error',
-     
-  },
-} satisfies Linter.Config as CompatibleConfig;
-
-const recommendedConfigMini = {
-  ...baseConfig,
-  rules: {
-     
-    'zod/consistent-schema-var-name': 'error',
-    'zod/no-any-schema': 'error',
-    'zod/prefer-meta': 'error',
-    'zod/require-brand-type-parameter': 'error',
-    'zod/require-error-message': 'error',
-     
   },
 } satisfies Linter.Config as CompatibleConfig;
 
@@ -139,7 +122,6 @@ export default {
   ...eslintPluginZod,
   configs: {
     recommended: recommendedConfig,
-    recommendedMini: recommendedConfigMini,
   },
 } satisfies ESLint.Plugin;
 /**
