+++ plugins/eslint-plugin-zod/src/rules/prefer-string-schema-with-trim.ts
@@ -1,4 +1,7 @@
-import { createZodSchemaImportTrack, findParentSchemaMatchingCondition  } from '@eslint-zod/utils';
+import {
+  createZodSchemaImportTrack,
+  findParentSchemaMatchingCondition,
+} from '@eslint-zod/utils';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
