+++ plugins/eslint-plugin-zod/src/rules/prefer-trim-before-string-length-checks.ts
@@ -1,4 +1,7 @@
-import { createZodSchemaImportTrack, findParentSchemaMatchingCondition  } from '@eslint-zod/utils';
+import {
+  createZodSchemaImportTrack,
+  findParentSchemaMatchingCondition,
+} from '@eslint-zod/utils';
 import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
