+++ plugins/eslint-plugin-zod/src/utils/create-plugin-rule.ts
@@ -3,12 +3,6 @@ import { ESLintUtils } from '@typescript-eslint/utils';
 
 import { getRuleURL } from '../meta.js';
 
-export type ZodPluginRule = ESLintUtils.RuleWithMetaAndName<
-  [],
-  string,
-  { zodImportAllowedSource: ZodImportAllowedSource }
->;
-
 export const createZodPluginRule = ESLintUtils.RuleCreator<{
   zodImportAllowedSource: ZodImportAllowedSource;
 }>(getRuleURL);
