+++ pnpm-workspace.yaml
@@ -1,24 +1,26 @@
 packages:
-  - "plugins/*"
-  - "packages/*"
+  - 'plugins/*'
+  - 'packages/*'
 
 catalogs:
   deps:
     # typescript-eslint ecosystem (keep versions aligned)
-    "@typescript-eslint/utils": "^8.57.0"
+    '@typescript-eslint/utils': '^8.57.0'
 
   dev:
     # typescript-eslint ecosystem (keep versions aligned)
-    "@typescript-eslint/rule-tester": "8.59.2"
-    typescript-eslint: "8.59.2"
+    '@typescript-eslint/rule-tester': '8.59.2'
+    typescript-eslint: '8.59.2'
 
     # build & test toolchain
-    tsdown: "0.21.10"
-    vitest: "4.1.5"
+    tsdown: '0.21.10'
+    vitest: '4.1.5'
 
     # plugin dev utilities
-    dedent: "1.7.2"
-    eslint-doc-generator: "3.3.2"
+    '@marcalexiei/prettier-config': '1.1.4'
+    dedent: '1.7.2'
+    eslint-doc-generator: '3.3.2'
+    prettier: '3.8.3'
 
 allowBuilds:
   unrs-resolver: true
